webtester/llm_integration_api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
langchain>=0.1.0
python-dotenv>=1.0.0
google-generativeai>=0.4.0 # For Gemini support
redis>=4.5.0 # For rate limiting
prometheus-client>=0.16.0 # For Prometheus metrics
opentelemetry-api>=1.16.0 # For observability
opentelemetry-sdk>=1.16.0
langchain-core>=0.1.0